+++ src/engine/Util/Coroutine.ts
@@ -1,12 +1,8 @@
 import { Engine } from '../Engine';
-// import { Entity } from "../EntityComponentSystem";
-// import { Scene } from "../Scene";
-
-export type CoroutineEventType = 'postupdate' | 'preupdate';
 export type CoroutineGenerator = () => Generator<number | Promise<any>, void, number>;
 
 /**
- * Excalibur coroutine helper, returns a promise when complete.
+ * Excalibur coroutine helper, returns a promise when complete. Coroutines run before frame update.
  *
  * Each coroutine yield is 1 excalibur frame. Coroutines get passed the elapsed time our of yield. Coroutines
  * run internally on the excalibur clock.
@@ -38,8 +34,8 @@ export function coroutine(engine: Engine, coroutineGenerator: CoroutineGenerator
           // schedule value milliseconds from now
           engine.clock.schedule(loop, value || 0);
         }
-      } catch {
-        reject();
+      } catch (e) {
+        reject(e);
       }
     };
     loop(engine.clock.elapsed());// run first frame immediately